+++ packages/template-retail-react-app/app/components/item-variant/item-attributes.jsx
@@ -8,12 +8,24 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import {FormattedMessage, FormattedNumber} from 'react-intl'
-import {Flex, Stack, Text} from '@chakra-ui/react'
+import {Flex, Stack, Text, Box} from '@chakra-ui/react'
 import {useItemVariant} from '@salesforce/retail-react-app/app/components/item-variant'
 import PromoPopover from '@salesforce/retail-react-app/app/components/promo-popover'
 import {useCurrency} from '@salesforce/retail-react-app/app/hooks'
 import {useCurrentBasket} from '@salesforce/retail-react-app/app/hooks/use-current-basket'
-import {usePromotions} from '@salesforce/commerce-sdk-react'
+import {usePromotions, useProducts, useProduct} from '@salesforce/commerce-sdk-react'
+
+/*
+    TODOs:
+        - pull out strings for localization
+            - selected options
+            - qty
+        - remove console.logs
+        - revert default.js
+        - ensure each child of list should have a unique key prop
+        - figure out bug where if you add multiple bundles with different variant choices, 
+        they appear under 1 product line item for order history
+*/
 
 /**
  * In the context of a cart product item variant, this component renders a styled
@@ -21,6 +33,8 @@ import {usePromotions} from '@salesforce/commerce-sdk-react'
  */
 const ItemAttributes = ({includeQuantity, currency, ...props}) => {
     const variant = useItemVariant()
+    console.log('VARIANT IN ITEM ATTRIBUTES', variant) // TODO: remove
+
     const {data: basket} = useCurrentBasket()
     const {currency: activeCurrency} = useCurrency()
     const promotionIds = variant.priceAdjustments?.map((adj) => adj.promotionId) ?? []
@@ -50,6 +64,68 @@ const ItemAttributes = ({includeQuantity, currency, ...props}) => {
         }
     })
 
+    // bundle logic for cart/order history page
+    const productBundleIds = variant?.bundledProductItems?.map(({productId}) => productId).join(',') ?? ''
+    const {data: productBundleVariantData} = useProducts(
+        {
+            parameters: {
+                ids: productBundleIds,
+                allImages: false
+            }
+        },
+        {
+            enabled: Boolean(productBundleIds),
+            select: (result) => {
+                return result?.data?.map((item) => {
+                    const flatValues = item?.variationAttributes?.flatMap((item) => {
+                        return item.values.map((vals) => {
+                            return {
+                                ...vals,
+                                label: item.name
+                            }
+                        })
+                    })
+                    const quantity = variant?.bundledProductItems.find((childProduct) =>
+                        childProduct.productId === item.id
+                    )?.quantity;
+                    return {
+                        ...item,
+                        quantity,
+                        variationValues: [
+                            ...Object.keys(item?.variationValues).map((variation) => {
+                                const found = flatValues.find(
+                                    (obj) => obj?.value === item?.variationValues?.[variation]
+                                )
+                                return {
+                                    value: item?.variationValues?.[variation],
+                                    label: found?.label,
+                                    name: found?.name,
+                                }
+                            })
+                        ]
+                    }
+                })
+            }
+        }
+    )
+    console.log(`productBundleVariantData for quacker ${variant?.name}`, productBundleVariantData)
+
+
+    // bundle logic for wishlist
+    const {data: productBundleData} = useProduct(
+        {
+            parameters: {
+                id: variant?.id,
+                allImages: false
+            }
+        },
+        {
+            enabled: Boolean(variant?.type?.bundle && !productBundleIds)
+        }
+    )
+    // console.log('productBundleVariantData IN ITEM ATTRIBUTES', productBundleVariantData) // TODO: remove
+    console.log(`productBundleData for ${variant?.name}`, productBundleData)
+
     return (
         <Stack spacing={1.5} flex={1} {...props}>
             {variationValues?.map((variationValue) => (
@@ -57,15 +133,6 @@ const ItemAttributes = ({includeQuantity, currency, ...props}) => {
                     {variationValue.name}: {variationValue.value}
                 </Text>
             ))}
-
-            {/* TODO: remove/modify */}
-            {variant?.bundledProductItems?.map((item) => {
-                return(
-                    <Text lineHeight={1} color="gray.700" fontSize="sm">
-                        {item?.itemText}
-                    </Text>
-                )
-            })}
 
             {includeQuantity && (
                 <Text lineHeight={1} color="gray.700" fontSize="sm">
@@ -75,6 +142,45 @@ const ItemAttributes = ({includeQuantity, currency, ...props}) => {
                         id="item_attributes.label.quantity"
                     />
                 </Text>
+            )}
+
+            {productBundleData && !productBundleVariantData && (
+                <Box>
+                    {productBundleData?.bundledProducts.map(({product, quantity}) => 
+                        (
+                            <Box marginTop={2} key={product.id}>
+                                <Text fontSize="sm" color="gray.700" as="b">{product?.name}</Text>
+                                <Text fontSize="sm" color="gray.700">Qty: {quantity}</Text>
+                            </Box>
+                        )
+                    )}
+                </Box>
+            )}
+
+            {productBundleVariantData && (
+                <Box>
+                    <Text fontSize={15} marginTop={3} fontWeight={500}>
+                        Selected Options:
+                    </Text>
+                    {productBundleVariantData?.map(({variationValues, name: productName, quantity, id}) => {
+                        return (
+                            <Box marginTop={2} key={id}>
+                                <Text fontSize="sm" color="gray.700" as="b">{productName}</Text>
+                                <Text fontSize="sm" color="gray.700">Qty: {quantity}</Text>
+                                {variationValues?.map(({label, name}) => {
+                                    return (
+                                        <>
+                                            <Text
+                                                fontSize="sm"
+                                                color="gray.700"
+                                            >{`${label} : ${name}`}</Text>
+                                        </>
+                                    )
+                                })}
+                            </Box>
+                        )
+                    })}
+                </Box>
             )}
 
             {variant.priceAdjustments?.length > 0 && (
