+++ packages/template-retail-react-app/app/pages/cart/index.jsx
@@ -352,9 +352,13 @@ const Cart = () => {
                             <GridItem>
                                 <Stack spacing={4}>
                                     {basket.productItems?.map((productItem, idx) => {
+                                        // bundles can be different as child products have different selections
+                                        const uniqueKey = productItem.bundledProductItems
+                                        ? productItem.productId + productItem.bundledProductItems.map((item) => item.productId).join(',') 
+                                        : productItem.productId
                                         return (
                                             <ProductItem
-                                                key={productItem.productId}
+                                                key={uniqueKey}
                                                 index={idx}
                                                 secondaryActions={
                                                     <CartSecondaryButtonGroup
